Live demo in Basic Block tutorial – redirect to a post with a live block example (#75)
* Live demo in Basic Block tutorial – redirect to a post with a live block example 

The current Live Demo link redirects to the plugins page, which isn't very helpful to a new person trying to interact with the block. This PR updates the link so that it redirects the user to an actual live example.

## Caveats

The `landingPage` link contains a hardcoded post ID right now. It's not ideal, as a future WordPress release could ship with different starter content and break that redirection. Playground will need to provide a mechanism for redirecting to a specific post, regardless of its relative ID.

* Do not urlescape all the characters in the Blueprint hash

* Remove double domain in the URL

* Escape the { character

* Different URL escaping technique

* Escape quotes

* Double escape quotes

* Update the Minimal Block example (no build)

* Attach the blueprints using the special _playground/blueprint.json file